server: moving file
