+++ test/server/server_fuzz_test.cc
@@ -7,7 +7,7 @@
 #include "source/common/network/address_impl.h"
 #include "source/common/thread_local/thread_local_impl.h"
 #include "source/server/listener_hooks.h"
-#include "source/server/server.h"
+#include "source/server/server_base.h"
 
 #include "test/fuzz/fuzz_runner.h"
 #include "test/integration/server.h"
