+++ tools/code_format/config.yaml
@@ -162,7 +162,7 @@ paths:
     - source/server/overload_manager_impl.cc
     - source/server/config_validation/server.cc
     - source/server/admin/html/active_stats.js
-    - source/server/server.cc
+    - source/server/server_base.cc
     - source/server/hot_restarting_base.cc
     - source/server/hot_restart_impl.cc
     - source/server/ssl_context_manager.cc
